feat: create cuctom hook useCard
